Rename android_utils.
